backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
google-generativeai==0.3.2
beautifulsoup4==4.12.2
requests==2.31.0
playwright==1.40.0
PyPDF2==3.0.1
python-docx==1.1.0
aiofiles==23.2.1
Pillow==10.1.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0 